Pull in flash size while parsing for flash verification line
Flash size can vary with different hex flashing. This parsing now works
will any flash sizes.

I was adding in multiple verification lines to look for in
flash_results.txt, but this is not going to work as we see more and more
flash sizes. I originally thought it was different because my initial
dev was done on a PTH 328, and then the second procedure I updated was
an SMD 328. But it turns out that the flash size changes minimally, and
so the parsing needs to be able to handle any flash size.