example serial upload bash file
This is the bash file and self test hex I am using for testing the FJ.
I thought it would be a useful starting point for creating other serial
upload bash files.

Note, on your Pi, these need to live in the directory
/home/pi/SERIAL_UPLOAD/ to work with test.py This will keep them
separate and organized away from the bootloader or combined hex. (while
still allowing for unique naming of hex files).

We are used to calling "COM3" or "COM4" in windows avr dude, so the main
difference here from our usual batch files is the "/dev/tty/USB0". When
I only have one FTDI basic plugged into my Pi, this is the correct com
port. I haven't tried multiple FTDI basics plugged in at the same time,
but I imagine it might be trickier to know which one's which.