Serial Upload to reference on MAIN board.
Planning ot make a little "add on" board to drop on top of the MAIN
board. It will provide serial upload success/fail LEDs off of GPIO 20
and 21. The idea is to have castellated edge PTH pins to connect to the
bottom 4 pins of the 2x20 connector. We really only need GND and the two
GPs, but I figured 4 will provide better stability.