Added "-i 1000"
Added in a "bitbang delay" into the AVRDUDE call for fuse bits. This
slows down the communication speed - ensureing we are well below what
some fresh chips need (if it's operating on internal oscilator at 1MHz,
then we need it to be SUB 125KHz).

Looking at the logic analyzer I see it's programming fuse bits at about
370 Hz (yes HERTZ, not KILO Hertz). This is plenty slow.

Without the delay flag included in the flash programming command, the
communication seems to be at about 60-80 KHz.

more info:

http://www.nongnu.org/avrdude/user-manual/avrdude_4.html#Option-Descriptions

-i delay

For bitbang-type programmers, delay for approximately delay microseconds
between each bit state change. If the host system is very fast, or the
target runs off a slow clock (like a 32 kHz crystal, or the 128 kHz
internal RC oscillator), this can become necessary to satisfy the
requirement that the ISP clock frequency must not be higher than 1/4 of
the CPU clock frequency. This is implemented as a spin-loop delay to
allow even for very short delays. On Unix-style operating systems, the
spin loop is initially calibrated against a system timer, so the number
of microseconds might be rather realistic, assuming a constant system
load while AVRDUDE is running. On Win32 operating systems, a
preconfigured number of cycles per microsecond is assumed that might be
off a bit for very fast or very slow machines.