ADD ON boards for SPI logic level conversion
The bi-directional logic level converter has been working in my testing.
I have yet to get the TXB0104 chip to work, so I just want to try it
out. (supposedly it will work better).

Also, created a mini library with the special raspi header - for top
mounting - it has larger drills and annular rings to fit on top of a
populated header with fillets.