Option to "tap" logic level from target
Note, this required a modification to the ADAPTER as well. I am
utilizing pin 6 on the 1x6 pin header (jumper wire) to transfer VCC back
to the logic level converter.

I'm not sure why on the original design I didn't think to include VCC in
the comm lines to the target - I think it was because the raspi doesn't
have any ADCs to actually  verify it, so I didn't know if it was worth
it.