moved 1x6 header to allow for RA header
This should allow for the cable to flow directly downward from the
pi_grammer and sit with a lower profile. This should help prevent the
cable/header from getting knocked around during handling.

Also had to update the adapter - rotated header 180 degrees so it will
mate straight pin-to-pin on the 1x6 header on the speedySPI.