Serial upload stuff
Added in LOCK_BITS_PASS variable so that I can avoid attempting a serial
upload if there is any failures upstream.

Also, started checking a "temp" copy of readout from serial program, to
adetify if it's looking hopeful. If it's going to fail, then it usually
takes about 30 seconds to timeout, and can appear to the tech that the
pi_grammer has frozen. This helps avoid this illusion.

Also, this is still not killing the subprocess, and I'm not sure how to
do that properly. So even though I move on in the python module, the
serial upload bash file is still going and timeing out again and again
until it's totally done in about 30 seconds.