pi_serial_upload default settings and variables
I added a device and baud rate to the top of this script similar to the
pi_upload shell script. I changed the programmer command (-c) to default
to arduino instead of wiring, the default device to atmega328p, and the
baud to 115200. I also put notes in the script to pick a slower baud
rate for slower clock speesd ( <= 8mHz)