Re-creating pi_program on Raspi to avoid strange CR behavior
Note, I was seeing some errors being thrown in the raspi when I tried to run pi_program.sh - it was showing that the command "sleep 0.1\r" was an invalid value for sleep. I'm assuming that the "\r" is some annoying carrage return from editing the file on a windows machine.

Added comments around the device variable