on serial upload failure - added killall avrdude
When serial upload fails, it will take FOREVER for avrdude to attempt (like 10 times) to serial upload, and it will timeout on each of those attempts. Rather than waiting for that to finish (which looks like the pi is frozen to the user), I take a snapshot of the results and parse them a bit. I can tell that it's going south, and then kill it. Previously the "process.terminate()" wasn't working  - and I think it still doesn't do much. But the call to "sudo killall avrdude" really does the trick.

Also note, in the previous test.py, even though the python module would move on, the avrdude instance would continue. This would lock up the com port and cause more failures down the road.

Seems to be working pretty darn well now. wahoo!